+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## V2.1.0 Particle Update
+- Added a blood impact mist particles for projectile impacts, which will move towards the damage source (with some random variance).
+- Changed blood spray particles to be less over the top by default.
+- Made blood spatters slightly smaller.
+- Blood spray will go in the opposite direction of the damage (with some random variance).
+- Added a Common Config option for damage type debug messages.
+- Added Client Config options for blood spray friction and blood mist scale.
+- Added a Common Config option for blood spray amount and damage caps.
+- Blood sprays are now 3D similar to V1.3.4.
+- Blood mist particles will be double the size on a killing blow.
+- Silver Fish is now registered to bleed green by default.
+- Bugfix for issue: [#53](https://github.com/CravenCraft/BloodyBits/issues/53)
+
 ## V2.0.2 Datapack Structure Hotfix
 - Fixed an issue with the built-in datapack structure. Changes `entity_type` to `entity_types` to be compliant with 
     the 1.20.1 datapack structure.